requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4

# Core NLP and ML packages
nltk==3.8.1
numpy==1.24.3

# HTTP requests for Ollama API
requests==2.31.0

# Embedding and text processing
sentence-transformers==2.2.2
transformers==4.36.2
torch>=2.2.0

# Web interface options
gradio>=4.15.0
streamlit>=1.29.0

# Document processing
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2

# Utilities
psutil>=5.9.6
requests>=2.31.0
numpy>=1.24.3
pandas>=2.1.4

# Optional: For better performance
faiss-cpu>=1.7.4 
tqdm==4.65.0
colorama==0.4.6